requirements.txt
streamlit
opencv-python
scikit-learn
numpy
joblib
Pillow
fpdf
tensorflow
